// src/types/contract_days_added.d.ts
export type contract_days_added = {
  draft_rfp_released: number;
  approved_by_acb: number;
  rfp_released: number;
  proposal_received: number;
  tech_eval_comp: number;
  negotiation_comp: number;
  awarded: number;
};

// draft_rfp_released Int
// approved_by_acb    Int
// rfp_released       Int
// proposal_received  Int
// tech_eval_comp     Int
// negotiation_comp   Int
// awarded            Int